Maked changes in some methods
Maked changes in methods

+++ src/main/java/com/my/airportproject/service/AuthService.java
@@ -2,17 +2,16 @@ package com.my.airportproject.service;
 
 
 import com.my.airportproject.model.dto.roles.ChangeRoleDto;
-import com.my.airportproject.model.dto.user.AdminChangeSomeUsername;
+import com.my.airportproject.model.dto.user.AdminChangeSomeUsernameDto;
 import com.my.airportproject.model.dto.user.ChangeMyUsernameDto;
 import com.my.airportproject.model.dto.user.UserRegisterDto;
+import com.my.airportproject.model.entity.Role;
 import com.my.airportproject.model.entity.User;
 import com.my.airportproject.model.enums.EnumRoles;
 import com.my.airportproject.repository.UserRepository;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -44,56 +43,64 @@ public class AuthService {
                 userRegisterDto.getCompanyName()
         );
         this.userRepository.save(user);
-        this.roleService.setRole(user.getUsername());
+        this.roleService.setRole(user.getEmail());
     }
 
     public User getUserByUsername(String username) {
         return userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found!"));
+                .orElse(null);
     }
 
     public List<User> getAllUsers() {
         return this.userRepository.findAll();
     }
 
-    public User findByUsername(String username) {
-        return this.userRepository.findByUsername(username).orElse(null);
-    }
-
 
     public User getByEmail(String email) {
-        return this.userRepository.findByEmail(email).orElseThrow(()-> new UsernameNotFoundException("Unexciting user!"));
+        return this.userRepository.findByEmail(email).orElse(null);
     }
 
     public boolean checkCompany(String value) {
         return value.length() >= 1;
     }
 
-    public boolean findByRole(ChangeRoleDto dto) {
-        EnumRoles newRole = dto.getRole();
-        User user = this.userRepository.findByUsername(dto.getUsername()).orElseThrow(()-> new UsernameNotFoundException("Unexciting user!"));
-        EnumRoles currentRole = null;
+    public boolean changeRole(ChangeRoleDto dto) {
+        EnumRoles newRoleEnum = dto.getRole();
+        User user = this.userRepository.findByEmail(dto.getEmail()).orElse(null);
         if (user != null) {
-            currentRole = user.getRoles().get(0).getName();
+            Role currentRole = user.getRoles().get(0);
+            Role newRole = new Role(newRoleEnum);
+            return !currentRole.equals(newRole);
         }
-        return !newRole.equals(currentRole);
+        return false;
     }
 
-    public void deleteRow(Long id) {
-        User user = this.userRepository.findById(id).get();
-        this.userRepository.delete(user);
+    public boolean deleteRow(Long id) {
+        User user = this.userRepository.findById(id).orElse(null);
+        if (user != null) {
+            this.userRepository.delete(user);
+            return true;
+        }
+        return false;
     }
 
-    public void changeSomeUsername(AdminChangeSomeUsername dto) {
-          User  user=this.userRepository.findByUsername(dto.getEmail()).orElseThrow(()-> new UsernameNotFoundException("Unexciting user!"));
-        user.setUsername(dto.getNewUsername());
-        this.userRepository.save(user);
+    public boolean changeSomeUsername(AdminChangeSomeUsernameDto dto) {
+        User user = this.userRepository.findByUsername(dto.getUsername()).orElse(null);
+        if (user != null) {
+            user.setUsername(dto.getNewUsername());
+            this.userRepository.save(user);
+            return true;
+        }
+        return false;
     }
 
-    public User changeMyUsername(ChangeMyUsernameDto dto) {
-            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-           User user = this.userRepository.findByEmail(authentication.getName()).orElseThrow(()-> new UsernameNotFoundException("Unexciting user!"));
-        return  user;
+    public User changeMyUsername(String username, ChangeMyUsernameDto dto) {
+        User user = this.userRepository.findByEmail(username).orElse(null);
+        if (user != null) {
+            user.setUsername(dto.getNewUsername());
+            this.userRepository.save(user);
+        }
+        return user;
     }
 
 }
